+++ Libft/ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: hutricot <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/07 14:30:01 by hutricot          #+#    #+#             */
-/*   Updated: 2018/11/07 14:45:29 by hutricot         ###   ########.fr       */
+/*   Updated: 2018/11/12 14:54:50 by hutricot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strncpy(char *dest, const char *src, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (src[i] && i < n)
